Working on the jupyter scripts
