+++ hedtools/hed/tools/__init__.py
@@ -2,4 +2,6 @@ from hed.tools.col_dict import ColumnDict
 from hed.tools.key_map import KeyMap
 from hed.tools.key_template import KeyTemplate
 from hed.tools.sidecar_map import SidecarMap
-from hed.tools.map_utils import get_new_dataframe, generate_filename, get_file_list, get_columns_info
+from hed.tools.map_utils import get_columns_info, make_combined_dicts
+from hed.tools.io_utils import get_file_list, get_new_dataframe, generate_filename, make_file_dict, \
+    make_info_dataframe, make_key, remove_quotes, reorder_columns, separate_columns
