+++ hedtools/hed/tools/key_map.py
@@ -1,6 +1,7 @@
 import pandas as pd
 from hed.errors.exceptions import HedFileError
-from hed.tools.map_utils import get_new_dataframe, get_row_hash, remove_quotes, separate_columns
+from hed.tools.map_utils import get_row_hash
+from hed.tools.io_utils import get_new_dataframe, remove_quotes, separate_columns
 
 
 class KeyMap:
