+++ hedtools/tests/tools/test_sidecar_map.py
@@ -1,7 +1,7 @@
 import os
 import unittest
 from hed.tools.sidecar_map import SidecarMap
-from hed.tools.map_utils import get_new_dataframe
+from hed.tools.io_utils import get_new_dataframe
 from hed.errors.exceptions import HedFileError
